Testing out corner cases in json with ghost_mode, thus finding that bases with depends on targets that in turn have depends on base will cause infinite recursion.  A safety check is needed.  However, dependencies on "base" are unnecessary.
The dependency generation from "jsonmake playtest" on all available bases
yields ambiguity for stock and ghost_mode - exactly the intended behavior
from the state of the JSON at this point.  This shows that the dependency
tracking and duck typing are stabilizing.

TODO:  load a local jsonmake_commands.py if present for functions referenced
in the .json file.  This means that this build system is hyper-configurable.
In fact, all it does is track dependencies, but decoupling that from the action code
allows it to be easily customized to different purposes and platforms.

ADVANTAGE:  all this for only 336 lines of Python as of this commit.