Project renamed.
The plugin/handler system is being cleaned up so that anything more than basic
behavior for bases, stocks, and queues are handled in the queue.  The stock
and base objects are default now, and the hander, if any, is chosen in the config file.

It's better to keep this out of the tinymake.json file anyway!