My reasonable expectation that just looking for "cat confusion agency" should call Confuse-A-Cat for the job isn't succeeding.
Puzzle-A-Puma is muscling in on the domestic cat market, you see. My build engine sees it match as specializing in felines, and they've got a cannon too!  So both agencies show up, their assets get muddled, and sergeant can't pick out which staff to get to build the stage.

No stage-building, no magic tricks, no surrealism, just a cannon.

I'm so worried that this cat might still be moping! 😾😟...😂