Success in the Confuse-A-Cat test!  This cat is getting properly confused!
A great deal of set logic in Enqueue() made this happen.  Also, the "sergeant" has to explicitly provide staff, which in the deliberately complex and abstract/duck-typed YAML, is necessary for the stage to be built, so magic_tricks and surrealism get hit, thus providing sufficient confusion to any feline.  This is far afield of what we'd expect any Makefile to do, out of all the build targets from the simple line "yamake cat confusion agency".

The default functionality for anything related to targets on the filesystem is to check for existence.  This is based on the logic that mtime on a single file or directory isn't always the right approach to determine a build.  More explicit control in a plugin-driven system seems appropriate.

TODO:  the action tags with all their commands for system commands, git merges, file manipulation is now needed.